Add paginate views Add filters ransack in cars views
