+++ app/helpers/application_helper.rb
@@ -1,0 +1,2 @@
+module ApplicationHelper
+end
