+++ app/views/cars/_form.html.erb
@@ -55,7 +55,7 @@
         </div>
       </div>
       <div class='mt-2'>
-        <%= f.button :submit, class: 'btn btn-primary' %>
+        <%= f.button 'submit' %>
       </div>
     <% end %>
   </div>
