+++ app/views/kaminari/_first_page.html.erb
@@ -1,0 +1,11 @@
+<%# Link to the "First" page
+  - available local variables
+    url:           url to the first page
+    current_page:  a page object for the currently displayed page
+    total_pages:   total number of pages
+    per_page:      number of items to fetch per page
+    remote:        data-remote
+-%>
+<li class="page-item">
+  <%= link_to_unless current_page.first?, content_tag(:i, nil, class:"tf-icon bx bx-chevrons-left").html_safe, url, remote: remote, class: "page-link" %>
+</li>
