+++ app/views/kaminari/_prev_page.html.erb
@@ -1,0 +1,11 @@
+<%# Link to the "Previous" page
+  - available local variables
+    url:           url to the previous page
+    current_page:  a page object for the currently displayed page
+    total_pages:   total number of pages
+    per_page:      number of items to fetch per page
+    remote:        data-remote
+-%>
+<li class="page-item">
+  <%= link_to_unless current_page.first?, content_tag(:i, nil, class:"tf-icon bx bx-chevron-left").html_safe, url, rel: 'prev', remote: remote, class: "page-link" %>
+</li>
