+++ app/views/layouts/application.html.erb
@@ -171,9 +171,11 @@
             <!-- Content --> 
             <div class="container-xxl flex-grow-1 container-p-y">
 
-            <h4 class="fw-bold py-3 mb-4">
+
+            <h4 class=" mb-4">
               <%= yield :breadcrumb %>
             </h4>
+
 
             <!-- Layout Demo -->
             <%= yield %>
